#332: Moved the PassiveJumpDamageIncreaseParser into the model folder
 The file was not modified yet.
